chore(vscode): rename download-copy-sqlite-esbuild.js → download-copy-sqlite.js
