+++ binary/utils/bundle-binary.js
@@ -11,7 +11,7 @@ const { TARGET_TO_LANCEDB } = require("../utils/targets");
 const fs = require("fs");
 const {
   downloadSqlite,
-} = require("../../extensions/vscode/scripts/download-copy-sqlite-esbuild");
+} = require("../../extensions/vscode/scripts/download-copy-sqlite");
 const { fork } = require("child_process");
 
 async function downloadNodeSqlite(target, targetDir) {
