+++ extensions/vscode/scripts/download-copy-sqlite.js
