+++ extensions/vscode/scripts/prepackage.js
@@ -9,7 +9,7 @@ const {
   autodetectPlatformAndArch,
 } = require("../../../scripts/util/index");
 
-const { copySqlite } = require("./download-copy-sqlite-esbuild"); // remains for sqlite only
+const { copySqlite } = require("./download-copy-sqlite");
 const { generateAndCopyConfigYamlSchema } = require("./generate-copy-config");
 const { installAndCopyNodeModules } = require("./install-copy-nodemodule");
 const { npmInstall } = require("./npm-install");
